--- lib-src/libnyquist/nyquist/nyqsrc/audacityfnintdefs.h
@@ -1,1 +1,0 @@
-extern LVAL xlc_aud_do(void);
