--- lib-src/libnyquist/nyquist/nyqsrc/audacityfnintptrs.h
@@ -1,1 +1,0 @@
-  { "AUD-DO",  S, xlc_aud_do},
