+++ lib-src/libnyquist/nyquist/nyqsrc/localdefs.h
@@ -1,4 +1,3 @@
 /* include actual local file headers: */
 #include "sndfnintdefs.h"
 #include "seqfnintdefs.h"
-#include "audacityfnintdefs.h"
