+++ lib-src/libnyquist/nyquist/nyqsrc/localptrs.h
@@ -2,11 +2,8 @@
  *
  * CHANGE LOG
  * 28-Apr-03  rbd  Removed "include switches.h" -- already included
- * 12-Feb-18  jkc  Added "include audacityfnintptrs.h" -- for internal Audacity commands
  */
  
 /* extension to xlisp */
 #include "sndfnintptrs.h"
 #include "seqfnintptrs.h"
-#include "audacityfnintptrs.h"
-
