+++ src/effects/nyquist/Nyquist.cpp
@@ -3159,6 +3159,49 @@ static LVAL ngettext()
       wxGetTranslation(string1, string2, number).mb_str(wxConvUTF8));
 }
 
+/*--------------------Audacity Automation -------------------------*/
+/* These functions may later move to their own source file. */
+extern void * ExecForLisp( char * pIn );
+extern void * nyq_make_opaque_string( int size, unsigned char *src );
+
+
+void * nyq_make_opaque_string( int size, unsigned char *src ){
+    LVAL dst;
+    unsigned char * dstp;
+    dst = new_string((int)(size+2));
+    dstp = getstring(dst);
+
+    /* copy the source to the destination */
+    while (size-- > 0)
+        *dstp++ = *src++;
+    *dstp = '\0';
+
+    return (void*)dst;
+}
+
+/* xlc_aud_do -- interface to C routine aud_do */
+/**/
+LVAL xlc_aud_do(void)
+{
+// Based on string-trim...
+    unsigned char *leftp,*rightp;
+    LVAL src,dst;
+
+    /* get the string */
+    src = xlgastring();
+    xllastarg();
+
+    /* setup the string pointers */
+    leftp = getstring(src);
+    rightp = leftp + getslength(src) - 2;
+
+    // Go call my real function here...
+    dst = (LVAL)ExecForLisp( (char *)leftp );
+
+    /* return the new string */
+    return (dst);
+}
+
 static void RegisterFunctions()
 {
    // Add functions to XLisp.  Do this only once,
@@ -3170,8 +3213,9 @@ static void RegisterFunctions()
       static const FUNDEF functions[] = {
          { "_", SUBR, gettext },
          { "ngettext", SUBR, ngettext },
-      };
+         { "AUD-DO",  SUBR, xlc_aud_do },
+       };
 
-      xlbindfunctions( functions, sizeof(functions)/sizeof(*functions) );
+      xlbindfunctions( functions, WXSIZEOF( functions ) );
    }
 }
